Create lccIndex.html
The Legendries character creator is a webpage which allows the user to create a character.

It has a form where the user can enter numbers for their character's attribute values. Using these numbers, the webpage calculates the total XP and GP cost to build the character. It also calculates the check values of each attribute.

When the body of the webpage is clicked on, it calls the function "runTheNumbers();" which can be found in "lccScript.js"